Add UniqueLocationList and exceptions
A unique location list is added to ensure no duplicates
in locations. Exceptions were created to handle conflicts
when a duplicate location is given or a location is not
found.

Extra methods involving deletion, editing of locations were
made. They are foreseen to be necessary in the adding of future
features involving deletion and editing of locations.